+++ core/src/main/java/org/trade/core/persistent/dao/CandleRepository.java
@@ -5,7 +5,7 @@ import org.trade.core.dao.AspectRepository;
 import java.util.List;
 
 
-public interface CandleRepository extends AspectRepository<Candle, Integer>, CandleRepositoryCustom {
+public interface CandleRepository extends AspectRepository<Candle, Long>, CandleRepositoryCustom {
 
     List<Candle> findByContractAndBarSizeOrderByStartPeriodAsc(Contract contract, Integer barSize);
