+++ core/src/main/java/org/trade/core/persistent/dao/CodeTypeRepository.java
@@ -3,7 +3,7 @@ package org.trade.core.persistent.dao;
 import org.trade.core.dao.AspectRepository;
 
 
-public interface CodeTypeRepository extends AspectRepository<CodeType, Integer>, CodeTypeRepositoryCustom {
+public interface CodeTypeRepository extends AspectRepository<CodeType, Long>, CodeTypeRepositoryCustom {
 
     CodeType findByName(String name);
 }
