+++ core/src/main/java/org/trade/core/persistent/dao/ContractRepository.java
@@ -4,7 +4,7 @@ import org.trade.core.dao.AspectRepository;
 
 import java.util.Optional;
 
-public interface ContractRepository extends AspectRepository<Contract, Integer>, ContractRepositoryCustom {
+public interface ContractRepository extends AspectRepository<Contract, Long>, ContractRepositoryCustom {
 
     Optional<Contract> findBySymbol(String symbol);
 }
