+++ core/src/main/java/org/trade/core/persistent/dao/ContractRepositoryCustom.java
@@ -5,7 +5,7 @@ import java.util.List;
 
 public interface ContractRepositoryCustom {
 
-    ContractLite findContractLiteById(Integer id);
+    ContractLite findContractLiteById(Long id);
 
     List<Contract> findContractByUniqueKey(String SECType, String symbol, String exchange, String currency,
                                            ZonedDateTime expiryDate);
