+++ core/src/main/java/org/trade/core/persistent/dao/PortfolioRepository.java
@@ -3,7 +3,7 @@ package org.trade.core.persistent.dao;
 import org.trade.core.dao.AspectRepository;
 
 
-public interface PortfolioRepository extends AspectRepository<Portfolio, Integer>, PortfolioRepositoryCustom {
+public interface PortfolioRepository extends AspectRepository<Portfolio, Long>, PortfolioRepositoryCustom {
 
     Portfolio findByName(String name);
