+++ core/src/main/java/org/trade/core/persistent/dao/RuleRepository.java
@@ -3,6 +3,6 @@ package org.trade.core.persistent.dao;
 import org.trade.core.dao.AspectRepository;
 
 
-public interface RuleRepository extends AspectRepository<Rule, Integer>, RuleRepositoryCustom {
+public interface RuleRepository extends AspectRepository<Rule, Long>, RuleRepositoryCustom {
 
 }
