+++ core/src/main/java/org/trade/core/persistent/dao/TradePositionRepository.java
@@ -3,6 +3,6 @@ package org.trade.core.persistent.dao;
 import org.trade.core.dao.AspectRepository;
 
 
-public interface TradePositionRepository extends AspectRepository<TradePosition, Integer>, TradePositionRepositoryCustom {
+public interface TradePositionRepository extends AspectRepository<TradePosition, Long>, TradePositionRepositoryCustom {
 
 }
