+++ core/src/main/java/org/trade/core/persistent/dao/TradelogDetailRepository.java
@@ -3,6 +3,6 @@ package org.trade.core.persistent.dao;
 import org.trade.core.dao.AspectRepository;
 
 
-public interface TradelogDetailRepository extends AspectRepository<TradelogDetail, Integer>, TradelogDetailRepositoryCustom {
+public interface TradelogDetailRepository extends AspectRepository<TradelogDetail, Long>, TradelogDetailRepositoryCustom {
 
 }
