+++ core/src/main/java/org/trade/core/persistent/dao/TradestrategyRepository.java
@@ -3,7 +3,7 @@ package org.trade.core.persistent.dao;
 import org.trade.core.dao.AspectRepository;
 
 
-public interface TradestrategyRepository extends AspectRepository<Tradestrategy, Integer>, TradestrategyRepositoryCustom {
+public interface TradestrategyRepository extends AspectRepository<Tradestrategy, Long>, TradestrategyRepositoryCustom {
 
     Tradestrategy findByRequestId(Integer requestId);
 }
